+++ changelog.md
@@ -16,6 +16,7 @@
  * Don't proceed with game launch if FA path is invalid (#933, #959)
  * Hotfix an irclib exception popping up in some cases (#958)
  * Delay displaying of user's ladder game info until launch (#360, #912)
+ * Use the more compatible ANGLE renderer by default (#963)
 
 Contributors:
  - Wesmania